accessible/tests/mochitest/tree/test_invalidationlist.html

<!DOCTYPE html>
<html>
<head>
  <title>Test document hierarchy</title>
  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../role.js"></script>
  <script type="application/javascript"
          src="../states.js"></script>

  <script type="application/javascript">
  function doTest() {
    var tree =
      {SECTION: [
       { role: ROLE_PUSHBUTTON, name: "Hello" },
       { SECTION: [
        { TEXT: [ { role: ROLE_TEXT_LEAF, name: "Hello" } ] },
        { role: ROLE_TEXT_LEAF, name: "World" }
       ]}
      ]};
    testAccessibleTree("container", tree);
    dumpTree("container");

    SimpleTest.finish();
  }

  SimpleTest.waitForExplicitFinish();
  addA11yLoadEvent(doTest);
  </script>
</head>

<body>
  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=673757"
     title="Do not process invalidation list while tree is created">
    Mozilla Bug 673757
  </a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <div id="container">
    <div role="button" aria-labelledby="a"></div>
    <div>
      <span id="a">Hello</span><span id="b">World</span>
    </div>
  </div>
</body>
</html>
